backend/requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
paho-mqtt